INDY-2147: Update tests
